Skip failing pandera guide test (#19912)
## Summary & Motivation

Some dependency of pandera is causing breakage of one of our
docs_snippets tests. The same issue is afflicting other users of
pandera: https://github.com/unionai-oss/pandera/issues/1500

Since it is not clear what nth-order dependency of pandera needs to be
pinned, just skip this test and wait for the upstream pandera fix.