[docs] - Update Helm guide instructions (DOC-148) (#19913)
## Summary & Motivation

This PR updates the copy in one of the Helm deployment guides to be a
bit clearer. Previously, the copy made it seem like instructions were
missing from the guide. This update now specifies what the example repo
does, instead of seeming like there should be steps to follow.

I also updated some links to use `PyObject` instead of Markdown links to
point to the API docs.

## How I Tested These Changes

👀