bigquery freshness helper (#19756)
This PR introduces a standalone method to the dagster-bigquery package
that, given a single table and a project ID, returns the freshness of
each table.

I tried to keep the core function as simple as possible, and use public
bigquery APIs rather than rely on raw queries of properties. This method
feels a bit more blessed. A previous version that tried to retrieve for
multiple tables felt weird to me because (a) there's no super convenient
bulk table retrieval method in bigquery and (b) requires us to make a
decision around how to handle exceptions for the user when only some of
the tables are causing exceptions.

How I tested this
Added a test that creates an observable source asset that attaches
freshness as metadata to its observe result, and asserts the information
is correct on the resulting observation event.