[external-assets] Change build_assets_job interface to use `executable_assets` and `loadable_assets` (#19765)
## Summary & Motivation

`build_assets_job` (an internal API) is the codepath through which all
asset job construction flows. Currently, it accepts two lists of assets:

- `assets: Sequence[AssetsDefinition]`
- `source_assets: Sequence[Union[AssetsDefinition, SourceAsset]]`

This is quite confusing. In particular, `source_assets` is poorly named.
Not only does it contain `AssetsDefinition` (which gets internally
converted to `SourceAsset` inside the function) but it also contains
observable source assets that actually have nodes that will be executed.

This PR changes `build_assets_job` to accept:

- `executable_assets: Sequence[Union[AssetsDefinition, SourceAsset]]`
- `loadable_assets: Sequence[Union[AssetsDefinition, SourceAsset]]`

This makes the purpose of these two assets arrays more clear. It can be
considered an incremental step in the source asset -> external asset
refactor.

The PR leaves the internal logic of `build_assets_job` unchanged (though
it will be changed upstack), passing separate collections of
`AssetsDefinition` and `SourceAsset` further into the system per the
status quo.

## How I Tested These Changes

Existing test suite.

A few GQL snapshots are updated due to minor changes in the asset base
jobs in the GQL repo. Specifically, the unexecutable external asset is
now no longer included in any of the base jobs where previously it was
included in all of them (as all unpartitioned assets defs were). This is
not a function of change to the `build_assets_job` logic but rather to
the base job construction routine (which is one of the two
`build_assets_job` callsites).