test(dbt-cloud): configure poll timeout when testing run materializations (#19941)
## Summary & Motivation
Some sample failures:
-
https://buildkite.com/dagster/dagster-dagster/builds/76208#018dc844-c2e2-4fa4-8e2a-dba9a356b98e
-
https://buildkite.com/dagster/dagster-dagster/builds/76219#018dc85c-e4ff-4633-8ecc-b49f18e9c289
-
https://buildkite.com/dagster/dagster-dagster/builds/76229#018dc862-cc64-43c7-8ed8-14a863211853

Since we enabled `pytest-xdist`, some tests may run slower since
multiple workers are running. Here, the test fails since:

1. A poll timeout was reached, when it was not expected to, or 
2. A poll timeout was reached, but earlier than expected (and so all the
request mocks were not called)

We solve (1) by removing the poll timeout in that case, since it is
unnecessary. The test would run to completion without it.

For (2), we remove the unnecessary request mocks and ensure that the
poll timeout is reached immediately.

## How I Tested These Changes
pytest