Ignore tempdir cleanup warning in test_utils.py (#19943)
## Summary & Motivation

Calling `JobDefinition.execute_in_process()` in certain circumstances
can trigger the emission of a `ResourceWarning` concerning a tempdir
cleanup, which in turn causes test failure for tests that are wrapped
with an "error on warning" fixture.

I'm pretty sure the ultimate solution to this problem is fixing the
`LocalArtifactStorage` tempdir cleanup code for ephemeral
`DagsterInstance`, which seems to have a bad interaction with instance
weakrefs. I'm not sure how to do this immediately, so we are going to
just filter out this (practically harmless) warning for now.