[bugfix] ExternalAssetNode AssetExecutionType backcompat fix (#19978)
## Summary & Motivation

Prior to this PR, `execution_type` on `ExternalAssetNode` was defaulting
to `MATERIALIZATION` if it was not specified in the constructor. This
caused an issue with backcompat for older versions of Dagster that did
not specify execution type, as observable and source external asset
nodes were erroneously being assigned a `MATERIALIZABLE` execution type.

The PR also reverts a change for observable status where we were looking
at `execution_type` instead of `is_observable`. `is_observable` is the
correct long-term solution in any case.

This resulted in the auto-materialize daemon, which reads
`ExternalAssetNode.execution_type`, not functioning as expected in some
cases.

## How I Tested These Changes

@OwenKephart Ran branch against malfunctioning deployment, observed
corrected behavior.