add optional op concurrency awareness into the queued run coordinator (#19181)
## Summary & Motivation
We want to prevent situations where concurrency-blocked runs are
spinning, spamming the event log for progress. This changes makes it so
that a narrow set of runs (where all the root nodes are concurrency
blocked) do not get dequeued.

The op concurrency counter keeps track of:
1. all the slots and pending steps for all concurrency keys matching the
concurrency keys required by all queued runs
2. all in-progress runs with `STARTING` run status and the concurrency
keys for all their root nodes
3. all the root node concurrency keys for the currently queued runs
about to be launched

Philosophically, any queued run that is able to make progress on any
step will get dequeued (absent any other constraints).

Additionally, the run coordinator can take an optional config for an
offset, so that more runs (or fewer) can get dequeued than are available
slots.

Similarly, a timeout in seconds can get set to treat runs in the
`STARTED` state as having claims on all of its root nodes concurrency
slots, to avoid aggressive dequeuing in between the run worker starting
and the step attempting to claim a slot. The window is generally only a
second, so this is probably not a big issue to fine-tune.

One big notable change is that all created runs will store the
concurrency keys of its root nodes at run creation time (as a property
of the dagster run). This requires the plan_snapshot that is created
with the run to reflect any asset selection / subselection. I have found
that this is not always accurately done in tests, but I *believe* all
the execution paths do this correctly.

*I spot-checked this by running the root concurrency key calculations on
existing asset runs for some cloud users.

Fixes FOU-12

## How I Tested These Changes
BK