Enforce that multi-assets must contain at least one out or at least one spec
Summary:
If you create a multi-asset with no outs or specs, the op for that multi-asset is added to every asset job (!) This doesn't seem like a valid state since every multi-asset should materialize at least one asset, so disallow it.

Technically this could be viewed as a breaking change? But the resulting behavior seems so clearly a bug that it seems reasonable to error here to me.

Test Plan: BK