[external-assets] Make AssetGraph error when fetching missing asset properties (#19973)
## Summary & Motivation

The `CachingInstanceQueryer` and other components of the asset daemon
sometimes rely on the asset graph reporting unknown keys as source
assets. The asset graph is being refactored to no longer support this
and error when reading properties for unknown keys. This PR adjusts the
callsites that rely on reporting unknowns assets as source assets to
instead perform an additional existence check for the asset.

## How I Tested These Changes

Existing test suite (which already includes cases for unknown assets).