[bugfix] Prevent AssetsDefinitions with no keys from being selected in assets jobs (#20006)
## Summary & Motivation

Currently `AssetsDefinition` nodes with no keys are included in every
assets job in the repo. This is due to a bug in subsetting logic. This
PR fixes that bug and ensures that `AssetsDefinition` nodes with no keys
are not included in any assets jobs.

There was one fivetran test that implicitly relied on the bug-- this
test was deleted after consulation with @OwenKephart:

>the test that's failing is the case where you specify no tables that
the connector will sync, and purely rely on the runtime asset
materializations. I think we can consider this invalid behavior -- it's
unclear why anyone would ever do this as you won't have any
representation of this in the asset graph, it'd be very difficult to
even find a way to execute the underlying op (I can't think of any way
to do it, to be honest)... just deleting the case should be fine

## How I Tested These Changes

Existing test suite