[ui] Add right-click menus on groups, assets in the sidebar (#19979)
## Summary & Motivation

This PR brings the context menus to the assets and groups shown in the
left sidebar. You could previously access these using the "•••" icon on
the far right, but now you can also right click!

Because our context menus are provided by a React hook, and hooks can't
be used conditionally, I had to split `AssetSidebarNode` into a common
container, and create a separate "label" with the content for each type
of node. It's only 11 more lines of code but it looks like a lot changed
in the file...
 

![image](https://github.com/dagster-io/dagster/assets/1037212/5352ca36-136c-418f-8883-7ea7ff0920cb)


## How I Tested These Changes

I tested this by right clicking all three kinds of nodes in the asset
graph sidebar. I also tried navigating with the arrow keys.

---------

Co-authored-by: bengotow <bgotow@elementl.com>