[external-assets] Change error to warning for duplicate assets when constructing ExternalAssetGraph (#20062)
## Summary & Motivation

#19900 introduced a check in the construction of `ExternalAssetGraph`
that would throw an error if two materializable or observable defs
having the same key were detected in a deployment (this is an invalid
state). However, after running a shadow script against customer
deployments, I discovered that ~3% of active deployments have this
problem, so I am changing it to a warning (and making the message more
informative) until we can find a clean strategy to message our customers
who have this problem.

## How I Tested These Changes

New test.