[ui] Upgrade yaml library, improve config merge behavior (#20055)
## Summary & Motivation

This PR aims to address #19668 (but doesn't entirely fix it):

Context: In the launchpad, there are several scenarios where we parse
the current YAML, modify it as a JS object tree, and serialize it back
to YAML. This is fundamentally undesirable because some YAML values (eg:
5.0) experience type coercion (eg: to 5) because JS does not have a
Float type, etc. Fully fixing this would require changing the YAML merge
logic to use YAML ASTs, but that's complex and we express config
defaults as JS objects, so it wouldn't entirely resolve the "5.0 to 5"
issue.

This PR fixes two of the more straightforward bugs in the "parse, merge,
stringify" workflow:

- The deepmerge library changed in v3+ to concatenate arrays by default
rather than combining them (so [a,b,c] + [c,d,e] = [a,b,c,c,d,e]). This
changes it to the expected ([a,b,c,d,e]), and means that merging two
identical arrays does nothing to them.

- The yaml stringify method tries to be "lean" with quotes, but can
un-quote some values that other YAML parsers parse as other types. The
example called out in #19668 was "5:30" becoming 5:30, which is a
Sexagesimal base-60 number in python. (This has been called out as awful
behavior before: https://news.ycombinator.com/item?id=26679229). I
changed the serialize call to explicitly maintain quotes on all string
values, which looks a bit ugly if you're a YAML purist but will avoid
these cases.

## How I Tested These Changes

I broke out a yamlMerge helper and wrote tests for the cases above,
verified they failed and then fixed them

---------

Co-authored-by: bengotow <bgotow@elementl.com>