[dagster-databricks] Fix setting databricks cluster node configuration
## Summary & Motivation
When submitting a Databricks Job, the DatabricksJobRunner is not setting the cluster's `driver_instance_pool_id` node configuration correctly. As a result, it is not possible to launch a Databricks Job from Dagster to run on Databricks instance pools. This sets the cluster's node configuration in accordance with the Databricks Jobs API to allow launching Databricks Jobs on instance pools.

## How I Tested These Changes
I launched Dagster runs in a Databricks workspace from local Dagster deployment.