[Asset checks] Fall back to selectedCheck.name (#20098)
## Summary & Motivation

Fix an issue where `selectedCheckName` is empty string because the user
didn't select a check and so we're falling back to the first check. In
this case we also need to fallback to the name of that check. The reason
there is both a selectedCheck and selectedCheckName in state is that
selectedCheckName (coming from URL parameter) determines the
selectedCheck, but when the URL parameter is absent the
selectedCheckName actually comes from the selectedCheck (the default
first one).

## How I Tested These Changes

`elementl cloud run shadow-dagit dogfood dogfood-test-1 --push-ssh-key
~/.ssh/id_rsa.pub --perms org-admin`

go to
https://dogfood-test-1.dogfood.dagster.cloud/prod/assets/test_prefix/checked_asset?view=checks