[docs] Hello, Dagster becomes Quickstart (#19981)
## Summary & Motivation

We would like to reduce the barrier to entry by replacing Hello, Dagster
with a Quickstart project. This project leverages GitHub Codespaces, so
that users can try Dagster without installing any dependencies on their
local machine.

This is a part of a larger initiative of simplifying the entrypoints
into Dagster and our documentation.

### Checklist

- [ ] Make the https://github.com/dagster-io/dagster-quickstart repo
public

### Outstanding Questions

- We have a few _quickstart_ templates that can already be scaffolded
with the `dagster` command. Where do we want to draw the line on this
quickstart vs those?

## How I Tested These Changes

- Ran Next.js project locally

---------

Co-authored-by: hyperlint-ai[bot] <154288675+hyperlint-ai[bot]@users.noreply.github.com>
Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>