[asset-backfill] Fix edge case when grouping runs with backfill policies and partition mappings. (#20110)
## Summary & Motivation

This PR refactors how we detect "simple" partition mappings, and limits
the scope of when we can group assets into runs in that case.

In terms of "simple" partition mappings, we now explicitly check for
Identity / TimeWindow partition mappings with no offsets. Previously, we
were looking for the case of "has a single upstream parent partition
that is the same as the child partition", which certainly sounds like a
"simple partition mapping" but actually does not cover all potential
cases.

In particular, imagine you have two daily-partitioned assets, with the
downstream asset depending on the previous N days of the upstream. In
this case, the first partition of the downstream asset only has a single
parent, and that partition is the same as the child partition, even
though this isn't a "simple" partition mapping! This meant it would be
erroneously allowed to run with its parent.

For the scope of when we group assets together, there were too many
weird cases to consider here and so instead I just reworked this logic
entirely to only allow weird partition mappings to run together if:

- The exact same set of upstream / downstream partitions are targeted
- All upstream partitions are set to be materialized on this tick

This means that we can be confident that all downstream partitions will
be requested this tick, and that they will be the same exact partitions
(meaning they will be grouped together into the same run)

## How I Tested These Changes

Test failed before, passes now