[RFC] Disable useQueryRefreshInterval background refresh while search is open (#20175)
## Summary & Motivation

Disabling the background refreshing while the search dialog is open so
as to limit the amount of background activity that could cause the
search experience to be laggy. Ideally I would love to allow the refresh
request to go through but pause the the actual React updates but to do
that I'd need to wrap `useQuery` which would require a lot of changes.
Curious whether y'all think I should just do that instead?

I think it makes sense to pause the background refreshing because
searching signals an intent to navigate away meaning the results of the
query **may** (may because it's possible they open in a new tab) not be
used.

## How I Tested These Changes

Tested with shadow-dagit with kraftheinz. This gets rid of all of the
smaller mainthread locks that happen while search is open except for one
caused by the very first fetch (we don't block the initial fetch) or
ones caused by fetches that were in progress when it was opened.