[bugfix] Fix copying of metadata, description in assets_def.from_attributes (#20189)
## Summary & Motivation

Copying of descriptions and metadata did not work correctly in
`with_attributes` (I discovered this in an upstack unrelated PR)--
instead of merging orig and replacements, replacements would totally
overwrite the originals. This only manifested as a problem for
graph-backed assets, I think because the original values could be
inferred in the constructor for non-GBAs.

## How I Tested These Changes

New test.