[dagster-databricks] Log databricks job run page url (#20089)
## Summary & Motivation
It is difficult to find the Databricks job run that is associated with a
Dagster run in the Databricks console. This logs an event that surfaces
the run page url for the launched Databricks job in the Dagster run
details page.

## How I Tested These Changes
Confirmed events were being logged in asset details page from local
deployment. The URL links to the current page if not databricks run page
url exists.

![Screenshot 2024-02-28 at 2 53
18 PM](https://github.com/dagster-io/dagster/assets/23409221/8d7297cb-40aa-4b22-9a93-0853f115a3ce)


Confirmed that the event link is omitted when a databricks run page url
does not exist.
![Screenshot 2024-02-28 at 5 37
12 PM](https://github.com/dagster-io/dagster/assets/23409221/704df78d-5547-460e-992a-97b019dbf2d3)