feat(dbt): log column schema only when using `DbtCliResource` (#20242)
## Summary & Motivation
In local development, when using the `dbt` CLI without `dagster-dbt`
(but with the `dagster` dbt package installed), seeing log messages with
column schema is distracting.

These log messages are only used by the `dagster-dbt` framework, and
should not be exposed to the user. So don't log them when just using
`dbt`, even if the `dagster` dbt package is installed.

## How I Tested These Changes
pytest