[external-assets] Hoist resolution of input asset keys to RepositoryDataBuilder (#20186)
## Summary & Motivation

We have some hairy logic that resolves "relative" asset keys associated
with asset inputs. Currently this resolution is done in multiple
internal places, which complicates internal codepaths dealing with
assets.

This PR hoists resolution to repository build time, in the same place
that we convert source assets to assets def. This simplifies internal
pathways.

In order to support this change, I had to alter tests that were directly
calling `UnresolvedAssetsJobDefinition.resolve` with an `AssetGraph` to
instead go through repository construction. With the external asset
conversion happening at the repo level, this is something that needed to
be done anyway.

## How I Tested These Changes

Existing test suite