feat(declarative-instigator-status): delete unloadable instigators in the scheduler daemon (#19988)
## Summary & Motivation
Delete unloadable instigators states managed by the scheduler daemon.

Note that logic was already in the schedule daemon, but not in the
sensor daemon. We explicitly do not add this logic in the sensor daemon
since accidentally deleting state (e.g. cursors) could be catastrophic.
In this case, explicitly add assertions to prevent sensor instigation
states from being deleted by the daemon.

## How I Tested These Changes
pytest