Fix duplicate events with asset check step launcher (#20239)
https://github.com/dagster-io/dagster/pull/19443 had a weak test case
that only ran a standalone step. It didn't test an actual run, which
would have failed with
`dagster._core.errors.DagsterInvariantViolationError: Expected to one
row for asset check evaluation, but updated 2.`

That diff was writing `AssetCheckEvaluationPlanned` events to the
external instance. However, step launchers log their events back to the
main instance so we ended up with 2 planned events. Duplicate planned
events meant that we create duplicate rows in the
`AssetCheckExecutionsTable` summary table. Options to avoid that
duplicate row are:

- [this diff] Don't log the planned events to the external instance. As
a result, we have to loosen a guarantee that every
`AssetCheckEvaluation` has a corresponding row in
`AssetCheckExecutionsTable`, created by the planned event. I'm not
thrilled about this, but I've been expecting to have to do the same
thing anyway to support wiping asset check executions (since you could
wipe between the planned event and the result, and we'd delete the rows
of the summary table)
- we could change `store_asset_check_event` to not deal with the summary
table on external instances. We'd need a toggle on the instance to know
if we're in an external context
- filter the `AssetCheckEvaluationPlanned` events out when we ingest the
external step launcher events back in to the main instance.
Unfortunately, step launchers call
`step_context.instance.handle_new_event(event)` directly inside of
`launch_step` instead of just yielding events out. We could change our
step launchers but I expect that user step launchers do the same