[external-assets] Standardize terminology on "execution set" (#20140)
## Summary & Motivation

The terminology around groups of assets/checks that must be executed
together is currently a little mixed:

- `AssetGraph` has `get_required_multi_asset_keys` and
`get_required_asset_and_check_keys`
- `ExternalAssetNode` and `ExternalAssetCheck` have a field
`atomic_execution_unit_id`

It's not obvious that these are related to one another. This PR attempts
to tighten up and standardize the terminology around the phrase
"execution set":

- `AssetGraph.get_required_multi_asset_keys` ->
`get_execution_set_asset_keys`
- `AssetGraph.get_required_asset_and_check_keys` ->
`get_execution_set_asset_and_check_keys`
- `ExternalAsset{Node,Check}.atomic_execution_unit_id` ->
`execution_set_id`

It also tweaks the return value of the `AssetGraph` functions-- the
current behavior is to return an empty set for subsettable assets, which
does not make sense for "execution sets"-- this changes it to return a
set with a single element (the passed asset key).

## How I Tested These Changes

Existing test suite.