[external-assets] Expose `asset_checks_defs_by_key` on `RepositoryDefinition` (#20222)
## Summary & Motivation

Currently `AssetChecksDefinition` are not stored directly on the
`RepositoryDefinition`-- they are only accessible indirectly via the job
defs they are associated with. This PR stores them side by side with
`AssetsDefinition` and `SourceAsset`. This is done both for consistency
and to support an upstack PR that makes an `InternalAssetGraph` the
source of "internal truth" for the repo.

## How I Tested These Changes

New unit test.