[AssetCondition] Refactor the testing suite (#20202)
## Summary & Motivation

This disentangles the generically-useful concept of "representation of a
set of definitions at a certain point in time" and that concept + an
instance from anything specific to AssetDaemonScenarios.

This allows these concepts to be reused in other places (the next target
for this is going to be testing AssetConditions in a vacuum, but this
same pattern would certainly be useful in backfill tests as well).


At a high level, AssetDaemonScenarioState is now a subclass of
ScenarioState, which is something that has a DagsterInstance + a
ScenarioSpec and provides convenience methods like `.with_runs()` etc.

A ScenarioSpec just contains information about definitions that you want
to exist, which is useful as it allows you to specify a base set of
assets / dependencies between them, then modify those structures (for
example by adding different partitions defintions). This enables the
creation of "libraries" of standard asset graph shapes (e.g. fan in, fan
out, hourly to daily, etc.), which multiple ScenarioStates can use.

## How I Tested These Changes