Fix bootstrapping behavior of MaterializeOnCron (#20020)
## Summary & Motivation

Previously, if you added this rule to your policy, it would not "notice"
that a cron tick had been missed until the first cron tick _after_ the
rule was added. This PR adds some additional logic to allow this asset
to detect that it's the first time this rule has been evaluated, and if
it is, does an explicit query to figure out if this asset has been
materialized since the previous cron tick.

Note that this implementation should get significantly simpler once we
port over to AssetConditions -- right now we basically have a replica of
the LatestPartitionsAssetCondition built into this rule. Ideally, we
could just refactor them both to call a single utility.

## How I Tested These Changes

Added test which failed before and passes now