Short-circuit in scheduler loop if we know not enough time has passed yet since the previous iteration
Summary:
The goal of this PR is to prevent us from spinning up a scheduler thread for each schedule in the workspace every minute, with the goal of improving scheduler latency for workspaces with large numbers of schedules.

Previously, every minute every schedule re-computed the next tick time, compared it to the current time, and short-cir. This is pretty wasteful for daily schedules (60*24 checks to re-compute the same data over and over again). Now, the initial thread keeps track of the result of its cron math for the next tick, and returns that timestamp to the main loop, where it is stored and used to short-circuit.

The timestamp isn't neccesarily the exact next tick, because we also periodically update the timestamp of each tick to make sure that if the cron schedule changes, we don't re-run anything that we didn't mean to (see https://github.com/dagster-io/dagster/pull/14561). We return the min of the next time we know we will plan to log the updated timestamp, and the cron math timestamp. This PR also makes that interval constant an env var so that we can tune it if needed for large workspaces.