Remove overly verbose logging in the scheduler and sensor daemon (#20274)
Summary:
Currently we log once a minute with more verbose information about what
each tick is doing. This can be useful to start but overwhelming with a
lot of schedules and sensors in the mix. add a flag to disable the
verbose logging altogether.

Test Plan: dagster dev with this env var set and unset

## Summary & Motivation

## How I Tested These Changes