fix(dbt): read dbt `+meta` from `+config` when present (#20296)
## Summary & Motivation
From https://docs.getdbt.com/reference/resource-configs/meta:

> Depending on the resource you're configuring, meta may be available
within the config property, and/or as a top-level key. (For backwards
compatibility, meta is often (but not always) supported as a top-level
key, though without the capabilities of config inheritance.)

This affects sources, as `+meta` config inheritance is applied in the
top level `+config` field, rather than in the top level `+meta` field.

This change reads from `+config` when present to read `+meta`.
Otherwise, we fall back to reading from the top level `+meta` field.

## How I Tested These Changes
pytest