[external-assets] Normalize assets defs in InternalAssetGraph (#20223)
## Summary & Motivation

Move the site of normalization of the assets definitions in a repo to
the `InternalAssetGraph`. Normalization is: source asset conversion +
dependency relative asset key resolution + generation of missing assets
def for referenced assets.

A previous PR introduced conversion of source assets to external assets
at the `RepositoryDataBuilder` level. However, we couldn't completely
expunge `SourceAsset` here because repositories have a public API that
returns `SourceAsset`-- so we needed to keep around the `SourceAssets`
passed into the repo, while only passing their external asset
representation into our internals.

This external asset representation of `SourceAsset` was merged into the
`assets_defs_by_key` map for the repo. Though we have not heard any
reports of problems from this, this was technically a breaking change,
as `assets_defs_by_key` would now return not only the assets defs a user
provided to a repo, but also the external asset representation of source
assets.

This PR improves on this situation in several ways:

- It reverts the change to `RepositoryDefinition.assets_defs_by_key`.
After this PR, this method will return only the `AssetsDefinition`
passed in by the user. This is consistent with `source_assets_by_key`.
- It moves asset normalization logic into `AssetGraph`. The `AssetGraph`
cached on a `RepositoryDefinition` becomes the source of truth for us
internally-- internal pathways operate on the normalized set of assets
obtained from the `AssetGraph`. This creates a clean separation between
our public exposure of the definitions our `RepositoryDefinition` takes
in and their private internal representation.
- A bonus is that our tests can keep using `AssetGraph.from_assets`
instead of going through a repository in order to get access to
realistic behavior (as there is no normalization stage skipped when
going through `from_assets`).

Most of the changes in the PR are just changing internal callsites of
`RepositoryDefinition.assets_defs_by_key` to
`RepositoryDefinition.asset_graph.assets_defs`.

## How I Tested These Changes

Existing test suite.