[AssetCondition] Create AssetConditionScenarioState class (#20278)
## Summary & Motivation

As we create more AssetConditions, it is increasingly useful to be able
to test them in isolation, rather than as part of a larger policy. This
builds off of previous work to split out the old AssetDaemonScenarion
test framework into something more reusable. Most of this PR is just
shifting existing functions around between files / folders.

Also added in a few tests for `AssetCondition.missing()` to demonstrate
how this would be used.

## How I Tested These Changes