fix(dbt): do not raise error when calling `.stream()` on `dbt source freshness` (#20315)
## Summary & Motivation
Resolves https://github.com/dagster-io/dagster/issues/20262.

Invoking `dbt source freshness` and calling `.stream()` is failing
because we are trying to yield events for a dbt source. We don't want to
yield events for dbt sources since we aren't modeling their
computations, but instead, just the presence of their reference.
Alternatively, we could yield `AssetObservation` for these events, but
we can wait until someone explicitly asks for that.

## How I Tested These Changes
pytest